Support adding modules from git repositories
Aside from using the defaul/global index, it's handy to have an
ability to add modules from just a git repository. With a
'cfbs.json' file with the repository-specific index it can be
used as a source of modules.

Ticket: CFE-3771
Changelog: 'cfbs add' now supports adding modules from git repositories with index