Download modules from our mirror instead of cloning
Cloning the repositories with modules is not a good idea,
especially because the repositories can be deleted by their
owners or some other issues can happen. Instead, when a module is
added to the index, we archive it in our mirror and then use the
archive in `cfbs build`, verifying that the checksum matches the
expected value stored in cfbs-index/master/versions.json.

Ticket: CFE-3780
Changelog: 'cfbs build' now downloads module archives instead of cloning repositories