Moved a lot of logic out of `commands.py`
`commands.py` was getting quite big and unmaintainable.
Separated out most functions that are not commands.
The ones that change `./cfbs.json`, such as the logic
for add command, were moved to `index.py`, which
needs a name change at this point. The lower level
functions for cloning, extracting, copying etc. were
moved to a new file, `internal_file_management.py`.

Only moved code, and added imports and renamed as necessary.
Tried to keep the code exactly as before. No changes in
behavior.

This commit doesn't make so much sense on it's own, but I
think it takes me in the right direction, and will make the
subsequent commits a lot easier to understand.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>