Fixed inconsistency when handling local folders as modules
In all other cases, the contents of the step folder is just the
file(s) related to that module, not an extra subdirectory.
And paths are relative to this folder (inside it).

This was just an accident introduced when adding support for
local directories, the code reused was the same as for local
files and in the case of a directory it just happened to create
an extra subdirectory with the module name.

This does change the behavior slightly for locally added folders,
but the fix is easy:

```
directory ./ <destination>
```

Instead of:

```
directory ./<module_name> <destination>
```

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>