Added canonify utility function
This function is inspired by the `string_lib.c:CanonifyNameInPlace`
function from libntech:

```C
void CanonifyNameInPlace(char *s)
{
    for (; *s != '\0'; s++)
    {
        if (!isalnum((unsigned char) *s))
        {
            *s = '_';
        }
    }
}
```

I tested it with the CFEngine `canonify` function to see if it produces
the same results. But it responds a little bit differently when using
non-ASCII characters. E.g. the emoji 🐿️ is translated to `___` by
CFEngine but `_` by this python function.

One thing to note is that the `isalnum` function responds differently
based on what locale is set on the system.

Ticket: None
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>