Add --masterfiles option for cfbs init
You now can specify version of masterfiles in cfbs init using the
--masterfiles option. The --masterfiles option can be set to `no`,
resulting in no masterfiles added. It can be set to a version number
matching regex `[0-9]+(\.[0-9]+){2}(\-[0-9]+)?` which adds a specific
version of masterfiles. Else the argument is interpreted as a branch
name, resulting in cfbs trying to add masterfiles using that branch.

Ticket: CFE-4071
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>