Fixed bug where repo is cloned twice
Fixed bug where cfbs attempts to clone repository twice if there is a
typo in the subdirectory key in cfbs.json. cfbs now terminates with a
descriptive error message.

Ticket: CFE-4110
Changelog: Fixed bug where cfbs attempts to clone repository twice if there is a typo in the subdirectory key in cfbs.json.
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>