Fixed issues when adding modules to missing build key
"build" key is only required for the "policy-set" type.
However, we still want users to be able to use the "build"
key and "cfbs build" in all projects, so we need to add the
key for them when necessary, and account for the fact that
the key might be missing.

Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>