GH Actions: Added build all test from build-index repo
This test might run into problems like being too slow, or running
out of API quotas / rate limiting. We might disable it in the
future, or merge it into the other similar test, at
'tests/shell/validate.sh'.

However, for now, it is useful, especially as I'm debugging why this
fails in the other repo.

Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>