Fixed repeated warnings in cfbs validate while returning 0 (success)
When we use warn_about_unused_keys from _validate_config() we really
want it to raise exceptions, like the rest of the valiation code.
validate_config() will handle the exception, and the code calling it
will determine how strict to be when it comes to the result.

(cfbs validate should be very strict, cfbs build is intentionally not
strict for backwards compatibility, other commands are somewhere in
between).

This fixes an issue seen in the build index tests, where these warnings
are printed many times by cfbs validate, because the "index" field is
not recognized, and cfbs validate still exits with 0 (success).

Further fixes are needed for other commands and for making cfbs validate
recognize the "index" field, this is just the first step.

Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>