Made validation more strict in cfbs add/remove/update/status/input
We want cfbs build to be backwards compatible (also cfbs download).
In these other commands it makes more sense to be more strict to
try to catch mistakes and prevent bad situations.

If cfbs versions differ, we assume the user is running a newer version and
the hub is running an older version of cfbs when it's running cfbs build.
In all cases, we want users to be able to upgrade cfbs and still have
cfbs build work.

If the user attempts to edit the project, and validation fails, it
could mean:

- The project was made / edited on a newer version of cfbs
- The project has mistakes, for example after editing the JSON manually

Both of which we want to catch and fix, by making the user upgrade cfbs
and/or fix errors in the cfbs.json.

Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>